refactor: consolidate optimistic inventory rollback into ClientNetwork
Move the PendingActionTracker into ClientNetwork as a single shared
tracker. applyOptimisticRemoval() now auto-snapshots before mutation
and manages rollback internally. onInventoryUpdated clears the tracker
when the server confirms. Pruner interval starts lazily and cleans up
on disconnect.

This eliminates:
- Duplicate PendingActionTracker instances in InventoryActionDispatcher
  and InventoryInteractionSystem (two timers, two listeners, potential
  overlapping rollback race)
- Runtime import of PendingActionTracker from client/ path in
  InventoryInteractionSystem (which lives in shared/)
- ~70 lines of boilerplate tracker setup across both callers

Callers now just call network.applyOptimisticRemoval() — one line.
